+++ CHANGELOG.md
@@ -11,10 +11,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Deprecated context manager on `Streamer`. Instead, use the object directly.
 
-### Added
+### Added or enhanced
 
 - `mpservice.streamer.Parmapper.__init__` takes two new arguments `executor_initializer`
   and `executor_init_args`.
+- Simplifications to the implementation of `streamer.py`, making use of `GeneratorExit`.
 
 
 ## [0.11.8] - 2022-12-21